refactor: move ExpressionValidator to Expression
